Renamed transports to adapters
