+++ src/adapters.rs
